[fix] int -> long

+++ src/main/java/com/jocoos/mybeautip/domain/event/service/dao/EventDao.java
@@ -2,9 +2,6 @@ package com.jocoos.mybeautip.domain.event.service.dao;
 
 import com.jocoos.mybeautip.domain.event.code.EventStatus;
 import com.jocoos.mybeautip.domain.event.code.EventType;
-import com.jocoos.mybeautip.domain.event.persistence.domain.Event;
-import com.jocoos.mybeautip.domain.event.persistence.repository.EventRepository;
-import com.jocoos.mybeautip.domain.event.vo.EventSearchCondition;
 import com.jocoos.mybeautip.domain.event.dto.EventStatusResponse;
 import com.jocoos.mybeautip.domain.event.persistence.domain.Event;
 import com.jocoos.mybeautip.domain.event.persistence.repository.EventRepository;
@@ -18,7 +15,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.ZonedDateTime;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 import static com.jocoos.mybeautip.domain.event.code.EventStatus.PROGRESS;
 
@@ -48,7 +44,7 @@ public class EventDao {
     }
 
     @Transactional(readOnly = true)
-    public List<Event> summary(Long eventNum) {
+    public List<Event> summary(int eventNum) {
         EventSearchCondition condition = EventSearchCondition.builder()
                 .statuses(Collections.singleton(PROGRESS))
                 .between(ZonedDateTime.now())
